patient-care-analysis-main/requirements.txt
google-generativeai
streamlit
pymongo
fpdf
schedule
python-dotenv
scikit-learn
networkx
pandas
imblearn
streamlit-lottie
bcrypt